backend/requirements-dev.txt
-r requirements.txt

# Development & Testing
flake8==7.2.0
black==24.12.1
mypy==1.15.0
pytest==8.3.5
pytest-django==4.11.0
pytest-cov==6.0.0
factory-boy==3.3.1

# Debugging
django-debug-toolbar==5.0.0
ipython==8.31.0

# Security scanning
safety==3.2.14
bandit==1.8.0